+++ v2g-liberty/rootfs/root/appdaemon/apps/v2g_liberty/data_import/__init__.py
@@ -1,1 +1,0 @@
-"""FlexMeasures data importing module."""