+++ v2g-liberty/rootfs/root/appdaemon/tests/v2g_liberty/test_reservations_client.py
@@ -1,13 +1,27 @@
 """Unit test (pytest) for reservations_client module."""
 
-from unittest.mock import MagicMock, patch, AsyncMock
 import pytest
+from unittest.mock import MagicMock, patch, AsyncMock
+from datetime import datetime
 from apps.v2g_liberty.reservations_client import ReservationsClient
 from apps.v2g_liberty.event_bus import EventBus
+from apps.v2g_liberty.evs.electric_vehicle import ElectricVehicle
+
 
 @pytest.fixture
 def event_bus():
     return AsyncMock(spec=EventBus)
+
+
+@pytest.fixture
+def mock_ev():
+    ev = MagicMock(spec=ElectricVehicle)
+    ev.max_range_km = 300
+    ev.max_soc_percent = 80
+    ev.soc_system_limit_percent = 97
+    ev.min_soc_percent = 20
+    return ev
+
 
 # Mock the log_wrapper to avoid actual logging
 @pytest.fixture
@@ -19,7 +33,14 @@ def mock_log_wrapper():
         yield
 
 
-# Test cases
+# Mock get_local_now
+@pytest.fixture
+def mock_get_local_now():
+    with patch("apps.v2g_liberty.v2g_globals.get_local_now") as mock:
+        mock.return_value = datetime(2025, 1, 1, 12, 0, 0)  # Example fixed time
+        yield mock
+
+
 @pytest.mark.parametrize(
     "summary, description, expected_soc",
     [
@@ -39,24 +60,20 @@ def mock_log_wrapper():
     ],
 )
 def test_add_target_soc(
-    mock_log_wrapper, monkeypatch, summary, description, expected_soc
+    mock_log_wrapper,
+    monkeypatch,
+    mock_get_local_now,
+    event_bus,
+    mock_ev,
+    summary,
+    description,
+    expected_soc,
 ):
-    """Test the _add_target_soc method
-    Assumed defaults in constants.py
-    c.CAR_MAX_SOC_IN_PERCENT = 80
-    c.CAR_MAX_CAPACITY_IN_PERCENT = 97
-    c.CAR_MIN_SOC_IN_PERCENT = 20
-    """
-
+    """Test the _add_target_soc method"""
     # Arrange
     hass = MagicMock()
     reservations_client = ReservationsClient(hass, event_bus=event_bus)
-
-    # TODO: The ReservationsClient should not have a method to set/get_constant_range_in_km
-    # but it is the only way i could get this to work.
-    org_value = reservations_client.get_constant_range_in_km()
-    reservations_client.set_constant_range_in_km(300)
-
+    reservations_client.set_vehicle(mock_ev)
     v2g_event = {"summary": summary, "description": description}
 
     # Act
@@ -64,9 +81,6 @@ def test_add_target_soc(
 
     # Assert
     assert result["target_soc_percent"] == expected_soc, (
-        f"Test failed for summary: {summary}, description: {description}. Expected {expected_soc}, but got {result['target_soc_percent']}."
+        f"Test failed for summary: {summary}, description: {description}. "
+        f"Expected {expected_soc}, but got {result['target_soc_percent']}."
     )
-
-    # TODO: The ReservationsClient should not have a method set_constant_range_in_km
-    # but it is the only way i could get this to work.
-    reservations_client.set_constant_range_in_km(org_value)